pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Rishikesh",email = "rishikesh2511990@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-debug-toolbar (>=5.1.0,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-ajax-datatable (>=4.5.0,<5.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false